+++ __init__.py
@@ -1,25 +1,24 @@
 # author: erosdiffusionai@gmail.com
-from .nodes_compositor import Compositor
-from .Compositor2 import Compositor2
-from .CompositorConfig import CompositorConfig
-from .CompositorConfig2 import CompositorConfig2
+# from .nodes_compositor import Compositor
+# from .CompositorConfig import CompositorConfig
+from .Compositor3 import Compositor3
+from .CompositorConfig3 import CompositorConfig3
 from .Alignment import Alignment
 
-
 NODE_CLASS_MAPPINGS = {
-  "CompositorConfig": CompositorConfig,
-  "CompositorConfig2": CompositorConfig2,
-  "Compositor": Compositor,
-  "Compositor2": Compositor2,
-  "Alignment": Alignment,
+    # "Compositor": Compositor,
+    # "CompositorConfig": CompositorConfig,
+    "Compositor3": Compositor3,
+    "CompositorConfig3": CompositorConfig3,
+    "Alignment": Alignment,
 }
 
 NODE_DISPLAY_NAME_MAPPINGS = {
-  "CompositorConfig": "CompositorConfig",
-  "CompositorConfig2": "CompositorConfig2",
-  "Compositor": "Compositor",
-  "Compositor2": "Compositor2",
-  "Alignment": "Alignment",
+    #"Compositor": "Compositor",
+    #"CompositorConfig": "CompositorConfig",
+    "Compositor3": "Compositor V3",
+    "CompositorConfig3": "Compositor Config V3",
+    "Alignment": "Alignment",
 }
 
 EXTENSION_NAME = "Enrico"
