+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "comfyui-enricos-nodes"
 description = "pass up to 8 images and visually place, rotate and scale them to build the perfect composition. group move and group rescale. remember their position and scaling value across generations to easy swap images. use the buffer zone to to park an asset you don't want to use or easily reach transformations controls"
-version = "2.0.4"
+version = "3.0.0"
 
 license = {text = "MIT License"}
